# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API server for asset management system.",
        "title": "Asset Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:6500",
    "paths": {
...
        "/asset/type/create": {
            "post": {
                "description": "Create one Asset Type record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Type"
                ],
                "summary": "Create one Asset Type record",
                "parameters": [
                    {
                        "description": "Asset Type Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AssetTypes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralCreateResponseBody"
                        }
                    }
                }
            }
        },
        "/asset/type/list": {
            "post": {
                "description": "List Asset Types",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Type"
                ],
                "summary": "List Asset Types",
                "parameters": [
                    {
                        "description": "List Action Record Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AssetTypeListDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AssetTypeList"
                        }
                    }
                }
            }
        },
        "/asset/type/one/{id}": {
            "get": {
                "description": "Get one Asset Type record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Type"
                ],
                "summary": "Get one Asset Type record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AssetTypes"
                        }
                    }
                }
            }
        },
        "/asset/type/update": {
            "post": {
                "description": "Update one Asset Type record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Type"
                ],
                "summary": "Update one Asset Type record",
                "parameters": [
                    {
                        "description": "Asset Type Request Body for update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AssetTypes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateInactiveResponseBody"
                        }
                    }
                }
            }
        },
        "/asset/type/void/{id}": {
            "delete": {
                "description": "Delete one Asset Type record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Type"
                ],
                "summary": "Delete one Asset Type record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateInactiveResponseBody"
                        }
                    }
                }
            }
        },
...
        "/base/code-type/batch-upload": {
            "post": {
                "description": "Batch upload code type record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Type"
                ],
                "summary": "Batch upload code type record",
                "parameters": [
                    {
                        "description": "Code Type Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CodeTypes"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CodeTypes"
                            }
                        }
                    }
                }
            }
        },
        "/base/code-type/create": {
            "post": {
                "description": "Create one Code Type record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Type"
                ],
                "summary": "Create one Code Type record",
                "parameters": [
                    {
                        "description": "Code Type Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CodeTypes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralCreateResponseBody"
                        }
                    }
                }
            }
        },
        "/base/code-type/get-type/{type}": {
            "get": {
                "description": "Get one code type record by type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Type"
                ],
                "summary": "Get one code type record by type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CodeTypes"
                            }
                        }
                    }
                }
            }
        },
        "/base/code-type/list": {
            "post": {
                "description": "List Code Types",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Type"
                ],
                "summary": "List Code Types",
                "parameters": [
                    {
                        "description": "List Action Record Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CodeTypeListDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CodeTypeList"
                        }
                    }
                }
            }
        },
        "/base/code-type/one/{id}": {
            "get": {
                "description": "Get one code type record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Type"
                ],
                "summary": "Get one code type record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CodeTypes"
                        }
                    }
                }
            }
        },
        "/base/code-type/update": {
            "post": {
                "description": "Update one Code Type record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Type"
                ],
                "summary": "Update one Code Type record",
                "parameters": [
                    {
                        "description": "Code Type Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CodeTypes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateResponse"
                        }
                    }
                }
            }
        },
        "/base/code-type/void/{id}": {
            "delete": {
                "description": "Delete one code type record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Type"
                ],
                "summary": "Delete one code type record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateInactiveResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AssetTypeList": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "lists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AssetTypes"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.AssetTypeListDto": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CodeTypeList": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "lists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CodeTypes"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.CodeTypeListDto": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
...
        "dto.GeneralCreateResponseBody": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.GeneralUpdateInactiveResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.GeneralUpdateResponse"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.GeneralUpdateResponse": {
            "type": "object",
            "properties": {
                "MatchedCount": {
                    "type": "integer"
                },
                "ModifiedCount": {
                    "type": "integer"
                },
                "UpsertedCount": {
                    "type": "integer"
                },
                "UpsertedID": {
                    "type": "string"
                }
            }
        },
...
        "models.AssetTypes": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "depreciationRate": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "typeCode": {
                    "type": "string"
                },
                "typeName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.CodeTypes": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "valueCode": {
                    "type": "string"
                },
                "valueName": {
                    "type": "string"
                }
            }
        },
...
    }
}